+++ ChangeLog.md
@@ -3,6 +3,9 @@
 
  * Migrate from Travis to GitHub Actions for CI
  * Add verbose option, by default do not print mafft messages
+ * Make --annotation-table argument take a single filename (rather than
+   comma-delimited) and allow the option to be used multiple times. E.g.:
+        flutile aadiff --annotation-table foo.txt --annotation-table bar.txt
 
 0.13.2 [2021-12-16]
 -------------------